requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
roastedbyai==1.2.0
Flask==2.3.3